Bring in all my good stuff (#4)
* Hard code my test key, and send messages to Segment directly.

* License under MIT, use WebKit style, and add sublime project.

* Initial test app.  Not automated yet, and no validation of results.

* Ignore all objects, test builds, and sublime workspace files.

* Added exceptions, and validation of test exceptions to posting.

* CMake support, and enhanced testing.  Internal API changes for RAII.

As part of adding these changes, all uses of new and delete have
been removed, using pass-by-reference so that exception handling
based cleanup can operate properly.

Some testing was done with valgrind, and there are hooks here, but
note that libcurl built with openssl (the default) breaks valgrind
badly on Ubuntu Trusty (and probably other systems!).  Further libcurl
has its own memory leaks.  As a result valgrind testing is off
by default.

* More RAII -- the library and tests are now free of all leaks.

While here, add the boilperlate to the example file.

* Fix build error on MacOS clang.

* Initial swag at CircleCI 2.0.

* Add nohlman's single-file json.hpp to the tree.

* Use 3rd party simpler JSON library.

* Initial swag at API for pluggable HTTP handler.

* Add <memory> to get shared_ptr.

* We need CMake 3.1 for the CMake CXX_STANDARD stuff.

This will break builds on Trusty, and hence Circle CI.
We need to get folks using newer stuff.

* Incorporate feedback from Achille.

We use shared_ptr to return the response back, and fixed a few
other issues.

* Pluggable HTTP transport methods -- baked in.

This ships with a CURL based implementation.  A little CMAKE
and #ifdef glue will be supplied later so that the code can be
compiled without bringing libcurl in at all (provided another
transport provider is available.)

* Asynchronous processing.

This uses a background thread, and so posting events always succeeds
from the callers perspective.  A Callback API is available for users
to obtain success/failure (and we use this in the test suite!), since
there is no way to use exceptions meaningfully to report errors.

* Linux build fixes.

This includes finding the threads library stuff, and some C++ tweaks
to conform to the standards as they exist on Linux (macOS was more
lenient.)

* Batched delivery.

This implements full batched delivery.  Batches are sent depending on
several tunables - a time based value, a count based one, and a size based
one.  There is also explicit flushing.  Even single events will be sent
using the batch mechanism for now.  Default thresholds are 10 seconds, 250 events,
or 500 KB.

There is no support for adding Context or Integrations objects to the batch
at this time.  Doing that will require exposing a new API that allows consumers
to attach that to the Analytics object.

* Seperate HTTP into its own namespace.

* Analytics into its own namespace now.

* Significant API improvements.

This includes a vanilla Object API for passing around Context and
Integrations objects, and the ability to access all the relevant
fields for every API call.  It also arranges to populate the timestamp
and sentAt fields correctly (accounting for queueing) and offers a
convenience method for users to obtain the current time as a nicely
formatted ISO8601 string.

* Fix cmake to not force CPP, and add missing sources.

* Travis CI fixes.

This uess container based matrixes and selects modern compilers; our code
requires GCC 4.9 or better, or Clang 3.4 or better.  As a result we can now
successfully test with Travis CI.

* AppVeyor support.

This is initial work; Windows builds are still broken, and need more
effort.  Nonetheless, AppVeyor stuff is now here.  We need someone
from the organization to actually setup AppVeyor on the main trunk.

* Substantial fixes for Windows.

When we switched to start testing with AppVeyor, I found many problems.
There appears to be a subtle race condition somewhere in the Windows
stack, but adding a brief delay (10ms) clears it and keeps the queue
from getting stuck.

While here, we have added a native WinInet transport which is used by
default on Windows instead of CURL, and CMake support to select it,
as well as a switch to remove all HTTP transports (-DNO_DEFAULT_HTTP=ON).

* Populate default context.

This also adds library and OS version information to the context.
While where we fixed up the README a bit.

* Fixes some badges to make things clearer.

Remove the garrett build status badges once this merges to master.

* Namespace cleanups for HTTP.

Eliminate the HTTP prefix on members of the segment::http namespace.
Also, return a unique_ptr instead of a shared pointer for the
response objects.

While here fix the README slightly.

* Better types for the entire project.

We move away from shared_ptr in many cases, and we use pass-by-value,
constant references, and standard C++ move semantics to get cleaner
and simpler code.

The Event object has just become an alias for Object (which itself is
an alias for the JSON object); we've provided nice factory functions
for this thing for the various types.  We also offer up a PostEvent
API that lets consumers post events that they have constructed directly.

There are also APIs to allow consumers convenience methods to adjust
things like anonymous ID, integrations, etc. on the top-level object.